Azure-IaC-Utils/infra_util_requirements.txt
# Azure ML & Management Libraries
azure-ai-ml
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-storage
azure-mgmt-machinelearningservices

# Environment & Authentication
python-dotenv
msal
requests

# Azure Search
azure-search-documents